requirements.txt
streamlit==1.31.*
pandas==2.2.*
numpy==1.26.*
plotly==5.18.*
google-cloud-storage==2.14.*
python-dateutil==2.8.*
pytz==2024.1

# Development
pytest==7.4.*
black==23.12.*
